Add 빈 토큰 지워주는 함수, Split 파일 나누기

+++ sanahn/libft/ft_strlen.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yecsong <yecsong@student.42seoul.kr>       +#+  +:+       +#+        */
+/*   By: sanahn <sanahn@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/15 16:32:17 by yecsong           #+#    #+#             */
-/*   Updated: 2022/03/25 11:16:41 by yecsong          ###   ########.fr       */
+/*   Updated: 2022/11/18 17:15:49 by sanahn           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
